Map, Items and Equipment update
Updated Area to include up and down on the map display. Fixed room reset. Added a way to view items,rarity,equipment slot,etc. in a tabled overview in-game. Made it so changes to equipment slot are only updated on the player once the changes have been saved. Made rarity dynamic it can now be modified via in-game commands.